# docs/swagger.yaml
definitions:
...
  models.Achievement:
    properties:
      foto:
        type: string
      id:
        type: integer
      image_url:
        type: string
      kategori:
        enum:
        - International
        - National
        - Campus
        type: string
      link:
        type: string
      nama:
        type: string
      pencapaian:
        type: string
    required:
    - kategori
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /achievements:
    get:
      description: Retrieves all achievements from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Achievement'
            type: array
      summary: Get All Achievement
      tags:
      - Achievement
    post:
      consumes:
      - multipart/form-data
      description: Insert a Achievements and saves them to the database.
      parameters:
      - description: Nama peraih achievement
        in: formData
        name: nama
        required: true
        type: string
      - description: Pencapaian yang diraih
        in: formData
        name: pencapaian
        required: true
        type: string
      - description: link ke Achievementnya
        in: formData
        name: link
        required: true
        type: string
      - description: Kategori Achievement
        in: formData
        name: kategori
        required: true
        type: string
      - description: Foto peraih Achievement
        in: formData
        name: foto
        required: true
        type: file
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Achievement'
      summary: Insert a new Achievement
      tags:
      - Achievement
  /achievements/{id}:
    delete:
      description: Delete a Achievements and its thumbnail from the database and storage.
      parameters:
      - description: Achievements ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Achievements deleted successfully
          schema:
            type: string
      summary: Delete Achievements
      tags:
      - Achievement
    get:
      description: Retrieves achievement data by its ID.
      parameters:
      - description: achievement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get achievement By ID
      tags:
      - Achievement
    put:
      consumes:
      - multipart/form-data
      description: Edits a Achievements by its ID
      parameters:
      - description: Achievement ID
        in: path
        name: id
        required: true
        type: string
      - description: Nama peraih achievement
        in: formData
        name: nama
        required: true
        type: string
      - description: Pencapaian yang diraih
        in: formData
        name: pencapaian
        required: true
        type: string
      - description: link ke Achievementnya
        in: formData
        name: link
        required: true
        type: string
      - description: Kategori Achievement
        in: formData
        name: kategori
        required: true
        type: string
      - description: Foto peraih Achievement
        in: formData
        name: foto
        required: true
        type: file
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Achievement'
      summary: Edit Achievements
      tags:
      - Achievement
  /achievements/category/{category}:
    get:
      description: Retrieves Achievements data by its Category.
      parameters:
      - description: Achievements Category
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get All Achievements By Category
      tags:
      - Achievement
  /achievements/foto/{id}:
    get:
      description: Retrieves the image of an Achievement by its ID.
      parameters:
      - description: Achievement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get Foto Achievement
      tags:
      - Achievement
...
swagger: "2.0"